fix: Include /types in package exports
This adds `./types/*` to the `exports` in `package.json`.

I don't totally understand what this does but I ran into an issue when
trying to import the `CssTransitionsRegistry` like in the README example
for "TypeScript Usage" after configuring TypeScript to use `"bundler"`
for `"moduleResolution"`.

I enabled `"bundler"` `"moduleResolution"` after seeing a comment from
@gossi suggesting that as a fix for another issue. Apparently
`"bundler"` is also now set by the latest version of `@tsconfig/ember`
and so others will likely run into this as well.

This seems to also work when `"moduleResolution"` is not set to
`"bundler"` but again, I'm not really sure what this is doing and/or why
it's needed.

References

https://github.com/buschtoens/ember-link/issues/777#issuecomment-1649255712
https://github.com/tsconfig/bases/blob/53d80ed4a9ece800ffca9774c4e63cef3ce6ab38/bases/ember.json#L13